Add T-GCN/GCN-LSTM to the various demo indices (#1440)
This adds "time series" section to the demos on Read the Docs. The
T-GCN/GCN-LSTM/`GraphConvolutionLSTM` notebook is included in this section, and
the various tables/summaries are updated to point to both the section and the
notebook.

This also renames the `spatio-temporal` folder to the slightly more general
`time-series` name, since GCN-LSTM can work on any time series data, not just
ones where the graph structure captures spatial proximity/connection. A
potentially better name would be `sequences` or `sequential`, since the
algorithm can work on any sequences.

Rendered:

- Demo landing page (with time-series blurb and table):
  https://stellargraph--1440.org.readthedocs.build/en/1440/demos/index.html
- Read the Docs time-series summary:
  https://stellargraph--1440.org.readthedocs.build/en/1440/demos/time-series/index.html
- Notebook:
  https://stellargraph--1440.org.readthedocs.build/en/1440/demos/time-series/gcn-lstm-LA.html

See: #1427