Rearrange demos to have fewer folders (#1471)
Per #1428, having many subfolders makes the docs display ugly and harder to
understand.

This also renames the demos to have a consistent and globally unique filename,
like `<algorithm>-<extras>-<task>`. This makes it easy to find a demo when
looking within a file browser (such as the left hand panel of Jupyter lab)
because they're sorted by algorithm name. It also ensures that a bug report just
mentioning a filename is enough to work out exactly which demo and what task
they're doing.

This retains an existing placeholder for the GCN node classification demo, since
we've linked to that in various places as a "get started with StellarGraph"
demo. An alternative would be to just rely/hope that we update all those lines
(#1482). The placeholder has the same name as the original, and so there's no
renamed detected for
`demos/node-classification/gcn/gcn-cora-node-classification-example.ipynb` to
`demos/node-classification/gcn-node-classification.ipynb`. However, this is just
a rename, with some of the internal links updated.

Docs view:
https://stellargraph--1471.org.readthedocs.build/en/1471/demos/index.html#table-of-contents

See: #1428