Use newer versions of requirements installed by Read the Docs (#1537)
This upgrades us to use the current latest versions of each of the dependencies
specified by Read the Docs. This gets us any improvements they include.

We specify them as exact versions in our requirements file because our
documentation is effectively a "binary" published artefact, not a library. This
means that users only interact with the output. The benefits include ensuring
that our documentation always "looks right", because we won't be accidentally
changing versions, and reproducibility between builds (including between the
build on Read the Docs and the one on CI). Cons include not getting automatic
updates (this doesn't seem to problematic, because we're just publishing static
HTML/CSS/JS, so security risks are relatively minimal).

(NB. one shouldn't pin versions when publishing a library, to avoid forcing
users to have particular dependencies installed, but this doesn't apply, because
this isn't part of the main StellarGraph library artefact.)

This fixes the security advisory we have, which is for Pillow < 6.2.2.

See: #1543