Add a notebook capturing StellarGraph class resource usage (#1547)
In many PRs, we've recorded the time and/or memory usage of several real-world
datasets, such as #1220, #1284, #1291, #1535.

This was done in an ad-hoc way, passing around various versions of a notebook as
gists, e.g.
1. initial version:
   https://gist.github.com/huonw/630907bf02f4ff443bb3d83139bb32aa
2. version with better source/target memory usage:
   https://gist.github.com/kieranricardo/536776d95bfbc56e8ab709f1d261cb50
3. version with end-to-end memory usage via `tracemalloc`:
   https://gist.github.com/huonw/deb88a9179cfb588f9d5fff59955f56d

This PR expands the last version to something that's appropriate to add to the
repo, so that it's easily accessible.

The notebook records both time and both long term and peak memory usage for
constructing five graphs:

- Cora
- BlogCatalog3
- FB15k with and without edge types
- reddit (optional, if manually downloaded from
  http://snap.stanford.edu/graphsage/#datasets)

These have a variety of sizes and "shapes" (Cora and reddit are homogeneous,
BlogCatalog3 has multiple node types and edge types, FB15k has multiple edge
types), although there's no heterogeneous graph with node features.

The memory usage is measured with [the `tracemalloc` Python stdlib
module](https://docs.python.org/3/library/tracemalloc.html). This counts every
single allocation, including ones from NumPy. The notebook loads the data
freshly from disk when tracing allocations, and thus captures every piece of
memory that goes into creating the raw data and the `StellarGraph` object. After
creating the `StellarGraph`, the notebook cleans up the input data, leaving only
the graph instance. At various points, the peak memory usage and snapshots of
the individual allocations are recorded. This allows computing various metrics:

- memory usage of the memory controlled by the `StellarGraph` instance itself
  (`memory (graph)`)
- memory usage of the _new_ memory controlled by the instance, that isn't shared
  with the raw data (`memory (graph, not shared with data)`)
- peak memory usage during construction of the graph (`peak memory (graph)`)
- peak memory usage during construction, ignoring the base memory usage of the
  nodes/edges input data (`peak memory (graph, ignoring data)`)
- base memory usage of input data (`memory (data)`)
- as a check, peak memory usage when loading the input data (`peak memory
  (data)`), if this is equal to `peak memory (graph)` it likely means that the
  peak memory usage during graph construction is less than loading data, and so
  the `peak memory (graph)` is not longer meaningful

All of this more accurate that using a retroactive analysis like [`asizeof` from
`pympler`](https://pythonhosted.org/Pympler/asizeof.html#asizeof), because it
correctly counts allocations from NumPy arrays once and only once. Pympler can
miss some memory: for `asizeof(obj)`, suppose `obj` contains an array that is
computed as a slice of another one (so `array.flags.owndata == False`), but
doesn't contain the parent/base array (`array.base`). `asizeof` seems to count
`array` as 0, and thus a potentially-large allocation isn't being counted. (Gist
2 above was attempting to fix this by inserting `np.copy`s at appropriate
places, but this is a fragile/inaccurate battle.)

This notebook is designed to be used by StellarGraph contributors, and so is
data-dense without too much explanation, and is placed into a new
`demos/zzz-internal-developers` directory. It is still added to `demos` and our
RtD, so that it runs on CI and for consistency.

See: #1544