Allow multidimensional features in StellarGraph (#1561)
This builds on the construct-from-NumPy support in #1556 to allow node features
to be multidimensional. That is, instead of just a 2D matrix of size N × F,
corresponding to feature vectors of length F for each of the N nodes, a
`StellarGraph` object can now have a feature array of shape N × F1 × F2 × ...,
corresponding to feature tensors of shape F1 × F2 × ... for each node.

This is useful for multivariate time series, where the most natural modelling of
the data is a time_steps × num_variates matrix (or the transpose) for each node.

None of our models will likely do anything useful with higher rank feature
tensors, at the moment (#1527). They'll all throw errors because
`StellarGraph.check_graph_for_ml` calls `StellarGraph.node_feature_sizes` which
emits an error for node types with complicated shapes, for backwards
compatibility.

See: #1524