Allow overriding the buildkite queue used for tests and notebooks (#1249)
It is good to be able to run tests on a GPU occasionally. GPUs have some
behaviour differences to CPUs, such as different function behaviour (less
validation, usually) and less reproducible behaviour. Without testing, we can't
be sure that the library fails because of these.

Running tests on a GPU all the time would be very expensive. In lieu of that,
this PR allows the agent that tests run on to be customised. This means that we
can occasionally run a build using a agent with GPU support, with a Buildkite
scheduled build.

This build should set the following environment variables, to use an agent with
a GPU, and to validate that the GPU is being used (this PR adds a test that
validates that we're running on a GPU when requested, to be sure things are
behaving correctly).

```
STELLARGRAPH_MUST_USE_GPU="1"
TEST_QUEUE_OVERRIDE="g4dnxlarge"
```

Example build using a GPU:
https://buildkite.com/stellar/stellargraph-public/builds/4321 notably, all the
steps that run on a GPU instance are slightly faster than the equivalent non-GPU
build (https://buildkite.com/stellar/stellargraph-public/builds/4317), e.g. 7
minutes for testing, vs. 10-12 minutes.

See: #1227