Use spelling extension to spellcheck Sphinx docs (#1583)
This uses https://sphinxcontrib-spelling.readthedocs.io to spellcheck our
rendered documentation. This includes:

- any text in `.rst` files in `docs`
- doc strings for functions/classes/modules
- text in notebooks

There's numerous examples of misspelled words, which this PR fixes. There are
also many words that are correctly spelled but aren't recognised as correct
(e.g. "GCN", names, and en-us spellings). This PR includes a whitelist of words
in `docs/spelling_wordlist.txt` (the default path for the spelling extension),
this whitelist includes the above, and can be expanded as necessary.

This is checked on CI, and includes an annotation containing the misspelled
words, as well as guidance for how to add things to the whitelist. For example:
https://buildkite.com/stellar/stellargraph-public/builds/4677

(https://github.com/sphinx-contrib/spelling/pull/44 will make the display even
more helpful by providing more context around a misspelled word).

See: #1368