Implement RotE and RotH (#1539)
This implements the RotE and RotH algorithms from [2] (aka the RotationE and
RotationH algorithms from [1]). After implementing them as described in [1], I
realised that [1] was probably a very-preprint, and it's now a proper preprint
at [2]. The latter expands the algorithm slightly to include reflections in
addition to rotations, but still explicitly discusses the rotation-only model
(under a very slightly different name). The RotE/RotH algorithm performs
similarly (or better than) the full AttE/AttH algorithm.

1. Chami I, Wolf A, Sala F, Ré C. Low-Dimensional Knowledge Graph Embeddings via
   Hyperbolic Rotations. http://stanford.edu/~fredsala/Hazy-HyperbolicKGs.pdf
2. Chami I, Wolf A, Juan D-C, Sala F, Ravi S, Ré C. Low-Dimensional Hyperbolic
   Knowledge Graph Embeddings. http://arxiv.org/abs/2005.00545

RotE is effectively a generalisation/combination of RotatE and TransE, and so
future work could be combining them all to have a single shared/configurable
model (including emergent models like hyperbolic TransE).

As somewhat of an implementation detail, this adds a `utils.hyperbolic` module
with the various functions required for the hyperbolic RotH. This allows them to
have some basic tests, as well as potentially making it easier to implement
other hyperbolic models, e.g.:

- https://cs.stanford.edu/people/jure/pubs/hgcn-neurips19.pdf
- https://arxiv.org/abs/1910.12892

Related discussion of RotationH/RotH and RotationE/RotE, including a comparison
to TransE, RotatE, etc.:
https://nbviewer.jupyter.org/gist/huonw/52d585e571fde62e252c54fbe668faa3

Follow up: #1664 adding a demo.

See: #1540