Add support for edge weights to GraphSAGE sampling (#1667)
This expands GraphSAGE (undirected and directed) to support weighted sampling,
where edges with higher weights are taken proportionally more often.

For example, suppose there's there's 4 edges from node A:

| source | target | weight |
|--------|--------|--------|
| A      | B      | 0      |
| A      | C      | 1      |
| A      | D      | 2      |
| A      | D      | 3      |

An unweighed walk starting at A will choose each of the edges with equal
propability and so end up on B, C or D in proportion 1:1:2 (edge counts). A
weighted walk will choose the edges proportional to the weights, so end up on
the vertices in proportion 0:1:5 (sum of edge weight). (Worth specifically
highlighting: a weighted walk will never chose the A-B edge because it has
weight 0.)

See: #462