Run notebooks on GitHub Actions (#1692)
This runs all the basic Jupyter notebooks on Github Actions, in parallel. It
does this in the simplest way to get parallelism (and also, the simplest way to
get a "loop" in GitHub Actions configuration), with a large job matrix:
https://help.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idstrategymatrix

The matrix has to be manually specified, so a new script
`scripts/ci/ci_workflow_notebook_list.py` is added, that checks that the matrix
of notebooks matches the notebooks on disk.

This script is better if it is simple, and it's better if the configuration all
in one place in the `ci.yml` file. Thus, the notebooks that aren't run in the
default configuration (such as Neo4j notebooks #1694, and notebooks that fail
#818 #819) are still listed and then excluded with `exclude`:
https://help.github.com/en/actions/reference/workflow-syntax-for-github-actions#example-excluding-configurations-from-a-matrix

See: #1687