Run Neo4j notebooks on GitHub Actions (#1694)
This expands on #1692 to also run the notebooks that use Neo4j on our new GitHub
Actions CI.

It does this by using docker-compose to manually run the Neo4j image in the
background. This is a bit clunky, but is done in the interests of getting
something working now. In particular, I couldn't work out how to use the native
support for "services" for a job
https://help.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idservices
to build the image from a local dockerfile.

See: #1687