Add error annotations in black GitHub Actions action (#1691)
This uses the `::error file=...::...` command
https://help.github.com/en/actions/reference/workflow-commands-for-github-actions#setting-an-error-message
to add annotations onto each file that's misformatted. This is a more consistent
experience for black, matching our other linters. It is also arguably better and
easier to understand, because it involves less digging through logs.

In particular, files that have incorrect formatting now have a "Check" comment
added to the first line. This is visible as a summary report when clicking on
"details" of the black step (for instance,
https://github.com/stellargraph/stellargraph/pull/1691/checks?check_run_id=794668066
) and also directly attached to each file under the "Files changed" tab.


See: #1687