Deprecate ClusterGCN class in favour of GCN (#1725)
The `GCN` class can now be trained with a `ClusterNodeGenerator` directly, and
so there's no need for the `ClusterGCN` class. This PR deprecates `ClusterGCN`
in favour of `GCN`. This makes the Cluster-"GCN" experience more uniform, where
there's various models that supports training with both `FullBatchNodeGenerator`
and `ClusterNodeGenerator`.

It does this without changing the functionality of the `ClusterGCN` class,
although it does swap out the implementation, so that `ClusterGCN` inherits from
(and uses the implementation of) `GCN`.

The demos are updated to use `GCN`, and some documentation is adjusted to
reflect that `GCN`, `GAT` and `APPNP` can be trained in this Cluster-GCN manner.

See: #1723