Allow the sidebar ToC to be deeper (#1512)
This overrides our settings to allow the sidebar's summary table-of-contents to
be deeper. This makes navigating the demos easier, e.g. it's easier to jump to a
specific section of a single demo, or jump to a related one (e.g. from GCN to
GAT).

This relies on the recent 0.5.0 release of the `sphinx_rtd_theme`.

Example:
https://stellargraph--1512.org.readthedocs.build/en/1512/demos/node-classification/gcn-node-classification.html

(See #1512 for a screenshot.)

See: #1499