Fix full batch tests on Windows (#1704)
Several tests for "full batch" methods like GCN, RGCN and APPNP manually
construct the tensor inputs, in the form of a sparse adjacency matrix, specified
as a tensor of indices and a tensor of data. The indices are created as a Keras
`Input` with dtype `int64`, but the model was being passed NumPy arrays with
dtype `int32`. This was apparently fine on macOS and Linux, but not on Windows
(likely due to the LLP64 behaviour of that platform, see #1705 for more
discussion).

This PR updates the tests to pass in NumPy arrays with dtype `int64`, and also
updates the RGCN generator to do the same conversion internally (matching the
non-RGCN full batch generator).

See: #1699