Allow easy unsupervised Cluster-GCN training with Deep Graph Infomax (#1257)
This does the small change required to allow for Cluster-GCN to be trained
easily with Deep Graph Infomax (in particular, enabling
`CorruptedGenerator(ClusterNodeGenerator(...))` to work without additional
parameters). It then also updates the demo indexing, _and_ updates the Deep
Graph Infomax notebook.

The updates to the notebook include adding a few more links to the API docs of
various algorithms, and also adding some headings for each algorithm, and
expanding the text. This is done here, because it's particularly helpful to be
able emphasise that GAT and APPNP can be trained using `ClusterNodeGenerator`
too.

See: #1256