Implement self-adversarial negative sampling knowledge graph loss (#1686)
This implements a slightly different loss function for knowledge graph
algorithms. It's so called "self-adversarial", because it uses the knowledge
graph's algorithm computed scores to give different weights to the negative
sampled edges. The intuition is that some edges end up being "obviously"
negative as the model trains, and so it's better to focus more attention on the
highly-scored negative edges and ignore the very low scoring ones.

Recall that knowledge graph algorithms typically work by taking in a set of true
edges, and, for each one, computing a series of corruptions or negative edges by
changing the source or target nodes (leaving the other node and the edge type
unchanged). This self-adversarial scoring works by grouping the negative edges
by their original "parent" edge, and computing a softmax over the scores within
that group. That is, for each true edge `(s, r, o)`, the scores of the negative
samples `(s', r, o)` and `(s, r, o')` are compared against each other to compute
the weights for that set of negative samples. Weights for negative edges created
from other true edges (even in the same batch) are computed independently.

This approach gives a small, but, apparently, noticeable improvement in the
performance of knowledge graph algorithms.

This is equations (5) and (6) from [1], and that paper includes an comparison of
different scoring techniques in Table 7 (screenshots below, for ease of
reference).

[1] Z. Sun, Z.-H. Deng, J.-Y. Nie, and J. Tang, “RotatE: Knowledge Graph
Embedding by Relational Rotation in Complex Space,”
<http://arxiv.org/abs/1902.10197>, Feb. 2019.

See: #1708