Add Actions job to serve as the 'required' pre-merge check on GitHub (#1732)
This PR adds a single CI Actions job `all-jobs-passed` that represents the
status of the whole CI workflow run. In particular, if the whole run passes, the
`all-jobs-passed` job will pass, and if any step fails, the `all-jobs-passed`
will be skipped and thus not run. This allows us to have a single status check
to mark as required, instead of having to mark every single testing and notebook
(and all the others) manually, in a way that cannot be validated automatically
(see #1739 for more discussion of this).

It works by having the `all-jobs-passed` step have a `needs` dependency
https://help.github.com/en/actions/reference/workflow-syntax-for-github-actions#jobsjob_idneeds
on every other step. This means that the `all-jobs-passed` step can only run if
all the other jobs pass.

Maintaining this list will be error prone, so the existing
`scripts/ci/ci_workflow_notebook_list.py` script is adapted to check this
`needs` list in addition to checking the notebook matrix. (The script was
renamed to `scripts/ci/validate_ci_workflow.py` and edited, but the changes are
significant enough that git is treating this as just a delete/add, not a
rename.)

Validation:

- all jobs passed, and so did the `all-jobs-passed` job:
  https://github.com/stellargraph/stellargraph/pull/1732/checks?check_run_id=817171302
- a job failed, and the `all-jobs-passed` job was skipped:
  https://github.com/stellargraph/stellargraph/pull/1732/checks?check_run_id=817050460

See: #1739