Compute notebook name for artifact upload before running (#1747)
We need to compute the output (name of the notebook) required for the artifact
upload step before running the notebook with Papermill. We use `set -e`, meaning
the script will stop at `papermill ...` if that invocation fails.

Example:
https://github.com/stellargraph/stellargraph/pull/1747/commits/bd2d7ce4db27adb3c7d08ab16f0408f650e09969
broke the GCN node classification notebook. `papermill` failed as expected, but
the notebook upload worked:
https://github.com/stellargraph/stellargraph/pull/1747/checks?check_run_id=817222719#step:7:11

See: #1746