Add neo4j version to notebook artifacts (#1753)
We run the Neo4j notebooks against both 3.5 and 4.0. The artifact upload
currently the same `name` for both versions, which means the artifacts overwrite
each other, and it's not possible to see the notebooks for both steps.

This PR adds the Neo4j version to the artifact name, to distinguish the two
different runs of each notebook.

Example: https://github.com/stellargraph/stellargraph/actions/runs/151111846
includes `undirected-graphsage-on-cora-neo4j-example_neo4j-4.0.ipynb` and
`undirected-graphsage-on-cora-neo4j-example_neo4j-3.5.ipynb`. See #1753 for a
screenshot.